refactor(executionIdUtil): distribute functions into appropriate modules and remove file
- Move getAncestorExecutionIds and getParentExecutionIds into nodeIdentification.ts:
  both are pure NodeExecutionId string operations with no external dependencies,
  consistent with the existing parse/create execution ID helpers in that file
- Move buildSubgraphExecutionPaths into workflowSchema.ts: it operates entirely on
  ComfyNode[] and SubgraphDefinition (both defined there), and isSubgraphDefinition
  is already co-located in the same file
- Delete executionIdUtil.ts and executionIdUtil.test.ts
- Distribute tests: ancestor/parent ID tests into nodeIdentification.test.ts,
  buildSubgraphExecutionPaths tests into workflowSchema.test.ts
